renamed SearchFilterWrapper to SearchFilters
